pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["slev-v <stasrudometkin2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
sqlalchemy = { extras = ["postgresql-asyncpg"], version = "^2.0.36" }
alembic = "^1.14.0"
uvicorn = { extras = ["standart"], version = "^0.32.1" }
psycopg2-binary = "^2.9.10"
httpx = "^0.28.0"
pytest-anyio = "^0.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
